passed swipe function to child instead of the entire deck data
